Restore old behavior of not matching on class owner
It turns out that invoking an interface method on a concrete class will
use an INVOKEVIRTUAL opcode with the concrete class name
rather than using an INVOKEINTERFACE with the interface class name.

To prevent an issue where stack map frames are generated improperly, only classes
compiled for Java 6 and above are supported. This shouldn't be an issue
in practice, since all mods should be targeting at least Java 6.